test: move poll_test_run_live tests into cli_test.py
console.py helpers exercised by a CLI command's behavior (like
output_merge_conflict_table) are tested in cli_test.py rather than a
standalone console_test.py, matching the existing BranchMergeConflictHelpersTest
and ChatLoopTest convention.
