+++ src/poly/tests/cli_test.py
@@ -744,6 +744,129 @@ class BranchMergeConflictHelpersTest(unittest.TestCase):
         self.assertEqual(len(table.rows), 1)
 
 
+class PollTestRunLiveTest(unittest.TestCase):
+    """Tests for poll_test_run_live error resilience (used by TestingCommand.testing_run)."""
+
+    @staticmethod
+    def _completed_run(test_id: str) -> dict:
+        """A fully-completed test run response for a single passing test."""
+        return {
+            "status": "completed",
+            "testHistory": [{"testCaseId": test_id, "status": "passed"}],
+            "passedCount": 1,
+            "failedCount": 0,
+            "errorCount": 0,
+            "testCaseCount": 1,
+        }
+
+    def setUp(self):
+        from collections import namedtuple
+
+        # Minimal stand-in for the TestCase objects poll_test_run_live expects:
+        # it only reads `.resource_id` and `.name`.
+        FakeTest = namedtuple("FakeTest", ["resource_id", "name"])
+        # All tests pass poll_interval=0 so the loop's time.sleep is a no-op
+        # and tests run instantly.
+        self.matched_tests = [FakeTest(resource_id="tc-1", name="My Test")]
+
+    def test_transient_errors_then_success_returns_completed_run(self):
+        """Recovers from a few consecutive transient errors and returns the run."""
+        import requests
+
+        from poly.output.console import poll_test_run_live
+
+        completed = self._completed_run("tc-1")
+        # First 2 polls fail transiently, the 3rd succeeds with a completed run.
+        get_test_run = MagicMock(
+            side_effect=[
+                requests.exceptions.HTTPError("500 Server Error"),
+                requests.exceptions.ConnectionError("connection reset"),
+                completed,
+            ]
+        )
+
+        result = poll_test_run_live(
+            get_test_run,
+            test_run_id="run-123",
+            matched_tests=self.matched_tests,
+            poll_interval=0,
+            max_consecutive_errors=5,
+        )
+
+        self.assertEqual(result, completed)
+        self.assertEqual(get_test_run.call_count, 3)
+
+    def test_error_counter_resets_after_successful_poll(self):
+        """A successful poll resets the counter so later errors are tolerated again."""
+        import requests
+
+        from poly.output.console import poll_test_run_live
+
+        completed = self._completed_run("tc-1")
+        # Pattern: fail, fail, succeed(pending), fail, fail, succeed(done).
+        # With max_consecutive_errors=3 this only completes if the counter
+        # resets after the pending success in the middle.
+        pending = {"status": "running", "testHistory": []}
+        get_test_run = MagicMock(
+            side_effect=[
+                requests.exceptions.HTTPError("boom"),
+                requests.exceptions.HTTPError("boom"),
+                pending,
+                requests.exceptions.HTTPError("boom"),
+                requests.exceptions.HTTPError("boom"),
+                completed,
+            ]
+        )
+
+        result = poll_test_run_live(
+            get_test_run,
+            test_run_id="run-123",
+            matched_tests=self.matched_tests,
+            poll_interval=0,
+            max_consecutive_errors=3,
+        )
+
+        self.assertEqual(result, completed)
+        self.assertEqual(get_test_run.call_count, 6)
+
+    def test_persistent_errors_give_up_and_return_empty_dict(self):
+        """After max_consecutive_errors failures it gives up without raising."""
+        import requests
+
+        from poly.output.console import poll_test_run_live
+
+        get_test_run = MagicMock(side_effect=requests.exceptions.ConnectionError("platform down"))
+
+        result = poll_test_run_live(
+            get_test_run,
+            test_run_id="run-123",
+            matched_tests=self.matched_tests,
+            poll_interval=0,
+            max_consecutive_errors=2,
+        )
+
+        # No successful poll ever happened, so the abandoned run returns {}.
+        self.assertEqual(result, {})
+        self.assertEqual(get_test_run.call_count, 2)
+
+    def test_happy_path_no_errors_returns_completed_run(self):
+        """A clean run with no errors returns the completed response on first poll."""
+        from poly.output.console import poll_test_run_live
+
+        completed = self._completed_run("tc-1")
+        get_test_run = MagicMock(return_value=completed)
+
+        result = poll_test_run_live(
+            get_test_run,
+            test_run_id="run-123",
+            matched_tests=self.matched_tests,
+            poll_interval=0,
+        )
+
+        self.assertEqual(result, completed)
+        self.assertEqual(get_test_run.call_count, 1)
+
+
 class ChatLoopTest(unittest.TestCase):
     """Tests for AgentStudioCLI._run_chat_loop.
 
@@ -2282,7 +2405,9 @@ class CreateProjectTest(unittest.TestCase):
     @patch("poly.cli_commands.project.AgentStudioInterface")
     @patch("questionary.text")
     @patch("questionary.select")
-    def test_user_cancels_project_name_entry(self, mock_select, mock_text, mock_iface_cls, mock_init):
+    def test_user_cancels_project_name_entry(
+        self, mock_select, mock_text, mock_iface_cls, mock_init
+    ):
         """create project returns early when user enters empty project name."""
         mock_iface = mock_iface_cls.return_value
         mock_iface.get_accessible_regions.return_value = ["us-1", "euw-1"]
